Moved everything non-templated to AbstractPluginManager, better plugin filenames.
Plugin filename prefix and suffix is now taken from CMake variables, so it's
fully multiplatform. But it requires generated header file (which contains
these variables set from CMake) and with current setup (where also
PluginManager.cpp is included from its header due to templates) both generated
header and cpp file must be shipped with library headers. It's ugly and it
requires fiddling with buildsystem to generate the header in EVERY project
depending on this library.

Now all templated code is in PluginManager.h, everything other is in
AbstractPluginManager.cpp.