PluginManager: fixed another static initialization order fiasco.
Static plugins imported with PLUGIN_REGISTER_STATIC cannot be added
directly to plugins() map, because at that time (i.e. before main() is
entered) we cannot be sure that all resources (plugin configuration
files) are registered. Because of that all information needed to
import these plugins is added to staticPlugins() vector and on first
call to plugins() static plugins are imported into the map.

Fixes for example issues with plugin names not visible for static
plugins, PluginTest is now passing again.