Debug output fix for Clang.
Clang has somehow broken template overload resolution and instead of
going up in the hierarchy and trying to find an alternative for
Debug::operator<< it tried to find an alternative for operator<< on
ostream and obviously failed.

Debug::operator<< is now moved out of class, which fixes the compilation
issue, but breaks tests, as Debug was copied into the function every
time and that broke this:

    Debug debug;
    debug << "a";
    debug << 33;

It wrote "a33\n" into the output, because it cannot notify original
instance to include whitespace before 33.

This is now fixed with new "weird" reference copy constructor, which
behaves the same as original copy constructor, but also notifies the
original instance to include whitespace next time. Hope this won't break
up again on other compilers.