PluginManager: removed metadata reloading and reload().
Automatic metadata reloading on load() or unload() could lead to
unexpected behavior when the plugin suddenly disappears from the list.
Better solution is to allow this to be done only explicitly by e.g.
reloadPluginDirectory(). Removed reload(), as the function would now
just unload the plugin and then load it again, which is meaningless.

Not yet sure how to properly implement reloading plugin metadata for
single plugins.