PluginManager: removed AUTOMATIC_INITIALIZER() from PLUGIN_REGISTER().
Static plugins are compiled into static library and because of that the
automatic initializer is either not called at all (because nothing
references the initializer and thus it is removed as unneeded) or called
twice (because PLUGIN_IMPORT() references it and also calls it). Now the
initialization is done only in PLUGIN_IMPORT().

Also changed the warning about static plugin added twice to assertion,
as it now shouldn't fire at all.