Utility: assert instead of error on not found resources.
It is programmer error not affected by any external state. Compiling
Utility library with CORRADE_GRACEFUL_ASSERT to be able to test the
assertions. To avoid compiling everything in Utility library twice, the
common part of both libraries is built as OBJECT target and then linked
together. This feature needs CMake 2.8.8.