TestSuite: allow explicit bool conversion with CORRADE_VERIFY().
Not doing explicit conversion with `bool()`, but rather passing the
value to if statement, which according to the standard is allowed to do
the conversion.

I could do the explicit conversion directly in the CORRADE_VERIFY()
macro and be done with that without all this verbose code, but this is
future-proof for the time we would want to ditch those macros
altogether.