More flexible ability to install CMake modules.
FindCorrade.cmake is not installed by default to avoid the need for
root privileges when installing, enable it with WITH_FIND_MODULE CMake
flag.

Other modules (UseCorrade.cmake and CorradeLibSuffix.cmake) are
installed to ${CMAKE_INSTALL_PREFIX}/share/cmake/Corrade for similar
reasons. FindCorrade.cmake is modified to find this path and add it to
CMAKE_MODULE_PATH, so the files can be conveniently included.