Revert "Utility: don't return after std::abort()."
With this change the compiler doesn't check for return value validity
anymore, which would lead to massive issues when using
CORRADE_GRACEFUL_ASSERT.

This isn't ideal, as the code for return value might be fairly large. I
need to find a way how to check the value and don't include it in the
ouptut unless CORRADE_GRACEFUL_ASSERT is used.

This reverts commit 7335a5e28ef0c55b1989f9adc8461d500af9a399.