PluginManager: removed "replacement" plugins.
Not implemented at all till today. The original usage was to provide the
ability to replace one plugin with another when the replacement cannot
be loaded alongside the original, but it is completely binary compatible
with the original one, so it can replace the original even as a
dependency. That is very constraining and has no real use in practice,
as for the binary compatibility it would need to have exactly the same
name and then we can just rewrite the original and not bother with any
replacements.