PluginManager: store ref to global storage per instance.
Avoids a nasty issue with static builds where the application was linked
to one instance of global plugin map and the dynamic loaded plugin was
linked to another instance of global map. The plugin registration then
asserted because the global instance owned by the plugin was empty and
the plugins seemed to be unknown.

This way the global plugin map instance is properly propagated to the
point of plugin registration. The only remaining issue with
duplication might be when the plugin itself creates its own
manager instance, but that's very uncommon.