Utility: get rid of annoying unsigned byte arrays.
I know, the `cha*` type is signed and that results in issues in some
particular cases. But in the majority of code we need to convert it to
`char*` and thus that results in annoying and ugly `reinterpret_cast`,
most annoying when writing simple test code.

In 90% cases this resulted in shorter lines and the cases where unsigned
type is really needed can be handled explicitly. Not doing any
compatibility/deprecation layer, as this was mostly in non-overloadable
return types and internal code. Sorry in advance if that affects
someone.