Containers: reworked Array constructors to be more explicit.
There are two new tags that distinguish the initialization operation.
Before the code looked like this:

    Containers::Array<int> a{5};
    auto b = Containers::Array<int>::zeroInitialized(5);

I hated the second line, because it was wayy longer than it should,
especially with long types. Also it wasn't exactly clear what kind of
initialization is done when you need to be really explicit. Now it is
like this:

    Containers::Array<int> a{5}; // eqivalent to the line below
    Containers::Array<int> a{Containers::DefaultInit, 5};
    Containers::Array<int> b{Containers::ValueInit, 5};

The new Containers/Tags.h header will be now used in more places to
facilitate different kinds of construction.