Interconnect: avoid casting to incompatible types.
It worked without issue on GCC and Clang, but casting a member
function pointer to completely different class that has absolutely no
relation to the other was crashing on MSVC (particularly the
slotInReceiverBase() test case), also completely broken in the
StateMachine test.

For the first time in history all Corrade tests pass on MSVC, yay!