Containers: custom deleters for Array.
Unlike std::unique_ptr, we can't afford the deleter to affect class
type. For example, in some image importer plugins, it is needed to call
std::free() instead of delete[], but the type
must still be Containers::Array<char>. Because of that, by default a
plain function pointer is used and the deleter is stored internally, but
it is possible to use different type if the deleter needs to be
stateful.