CMake subproject support, part 3: ability to override configure.h location.
The file is installed into the root include dir, but in the project it
is generated inside build tree and thus we need to override its location
for the FindCorrade.cmake module and then add additional directory into
the include path.

It means that the CORRADE_INCLUDE_DIR variable is no longer sufficient
and I had to expand it to CORRADE_INCLUDE_DIRS. As Corrade is almost
always used through Magnu, it should have no breaking effect on end
users, but as always, make sure you have up-to-date CMake modules in
your projects.