Containers: put back conversion constructors and operators.
I forgot to test properly with the other projects and it broke a valid
use case in Magnum, so the tests (and probably also some user code) were
not compiling. The valid use case is when one has an array of derived
type that is the same size as the base type so the element addressing
works properly (e.g. creating `ArrayView<Magnum::Math::Vector<3, Float>>`
from `Magnum::Color3*`).

This just puts back the original code and enables the same also in
StaticArrayView. It should be fixed properly with `std::enable_if`
checking also that both types have the same size, but that needs some
more thinking for properly functioning SFINAE (it gave compile errors
when using incomplete types) and this is an "build broken" emergency so
I'm leaving it as it is.

This reverts commit d4deecab787d02b3e5810473af06661a3e2450cd.