Utility: deprecate Debug::noNewlineAtTheEnd() in favor of Debug::Flags.
Compared to the previous usage:

    Debug::noNewlineAtTheEnd() << "Say hello:";

it's sadly slightly longer to write now:

    Debug{Debug::Flag::NoNewlineAtTheEnd} << "Say hello:";

but on the other hand much easier to implement and more extensible. I'm
sorry that I changed it so soon after the previous change (which was in
September, but it did not fit into upcoming colored output support.