TestSuite: ability to whitelist/blacklist test cases.
Using the --skip and --only command-line parameters the user can specify
which test cases to run. The test cases are denoted using numbers to
save typing (shh, actually I have implementation limit that prevents the
user to specify names directly). As everywhere else, multiple numbers
are separated with spaces and thus you have to enclose them in quotes.
With --only the order is preserved, allowing for user-defined ordering.

The test output now displays numbers next to test case names so it's
easy to see which test case has which number.