Containers: get rid of defaultDeleter() template.
Because the function was templated and the Array code used a pointer to
it, this meant the function had to be emitted *somewhere*. This caused
all hell to break loose especially in case of plugins that returned
Array instances -- if the plugin was unloaded before all Array instances
returned from it were destructed, the deleter was suddenly a dangling
function pointer, causing crash on Array destruction.

Now `nullptr` is used instead of a pointer to `defaultDeleter()` -- if
`nullptr` is encountered, the standard `operator delete[]` is called,
without any potential function pointer dispatch.

In general, plugins should be forbidden to use custom deleters
completely, this just fixes the most common case where default deleter
is used.