TestSuite: publicize testCaseRepeatId().
I originally didn't want to make this value public, as it would mean
that repeated test cases are not testing the exact same code repeatedly,
which was against the original intention of how should repeated test
cases be used. However now I found a new use case (testing 0-N
iterations of a floating-point algorithm and then verifying its
precision loss) and this is the exact thing I need -- instanced tests
would be too verbose if testing for e.g. 1000s iterations, but here it
will be quiet until the end or until it finds first failure, reporting
it for example as

    FAIL [01] iterative()@667

which would mean that the first 666 iterations passed but the 667th
failed.