TestSuite: huge documentation update.
Added a fully complete documentation of all features in the TestSuite
class (with images!) and updated the tutorial to include benchmarks as
well. The code snippets from TestSuite documentation were a bit involved
and I needed the output from the so they are now compiled as part of the
test run.