TestSuite: make the command-line arguments publicly available.
Makes implementation of derived tester classes (such as in Magnum) much
easier. There should be no user-facing change, though some formerly
publicly documented APIs were hidden from Doxygen as they should be used
only internally.

The arguments were changed to a mutable reference and non-const char
array pointer in order for more flexibility. Also, some third-party code
requires that (and most windowless apps in Magnum have it like that as
well).

In order to avoid forcing the users to pass argc/argv through to parent
constructor, the arguments are set via a static registerArguments()
function. The constructor asserts that the arguments are registered and
the destructor then wipes these registered arguments again in order to
avoid the argument registration being accidentally forgotten. However,
the arguments are available only through a member non-static function as
that is how it should be presented to the user.