Utility: better handling of boolean env values on Arguments.
Initial design assumed ON if the env variable was set (no matter what
value) and OFF if it wasn't. However, in TestSuite I have a
CORRADE_TEST_COLOR which is ON, OFF or AUTO. Some time ago I changed the
help text for boolean environment variables to say ON|OFF without
knowing that it *does not* behave like that. Now I realized that
inconsistency when writing the test case and finally made it all
consistent.