Utility: properly cleanup before exiting from Fatal.
I.e., insert a newline to the output and reset colors. The class is now
tested in a new FatalTest test, which is just a main() and CTest expects
that it exits with non-zero return code. Unfortunately it is not
possible to verify that the newline at the end gets inserted, as CMake
regexps can't handle multiline expressions (and they also then ignore
the return code and I WANT TO TEST BOTH DAMMIT).