TestSuite: add explicit test that verifies error code propagation.
This test, similarly to UtilityFatalTest, should return non-zero code
and thus signal the caller that it failed. If it returns zero (for
example the VM/emulator/interpreter/whatever running the test case is
not propagating error codes properly), it will be considered as failing.

This hopefully prevents future accidents where tests that should be
failing are silently ignored.