PluginManager: there is no need to have a global alias map.
Inter-manager plugin loading (for example when a font plugin depends on
an image loader plugin) is not using the aliases and that's the only
reason why there is a global plugin map in the first place. Made the
alias list a local thing, resulting in much simpler code (and hopefully
this gets me a bit further on the improved thread safety path).