PluginManager: remove non-reproducibility from the test.
Two plugins provided JustSomeMammal, so one of them got arbitrarily
picked up based on a phase of the moon (i.e., was working everywhere
except a single test case that was failing and only on Travis and only
sometimes). That's expected as it depends on whatever order the plugins
are found on the filesystem, so I'm adding a new alias that's unique and
does not have this problem.

Also documented the aliases a bit, as I'm getting a bit lost there.