Utility: implement "as documented" config subgroup shortcuts.
The problem was that

    [foo/bar]

was advertised in the Configuration documentation *since ever* (2010?),
but actually it never worked as that -- instead of expected group foo
with a bar subgroup, it caused a group named foo/bar to be parsed, which
is totally and completely wrong. That's now fixed. But, this was not
enough -- this exact thing should be also preserved on save, so not
expanding back into

    [foo]
    [foo/bar]

... which means, I also had to implement this shortcutting on the other
side. All in all, it basically took me the whole day because the parser
code had to be updated to a less crappy state first.