Utility: fix semantics of ConfigurationValue.
Before, a particular ConfigurationValue specialization was called only
if the input string was non-empty. That had two very bad properties:

 * It couldn't be used to parse types that don't have a default
   constructor
 * Configuration value parsers that wanted to return something else than
   the default-constructed value for an empty string had no chance to do
   it, as they got never called.
 * Moreover, the behavior of values() was different than value() here,
   making it really bad.

Changed that semantics to calling the ConfigurationValue specialization
always. That uncovered an uninitialized value bug in the default
ConfigurationValue parsers, so fixed that as well and extended their
documentation a bit.