PluginManager: make it possible to modify global plugin configuration.
Compared to modifying instance-specific configuration through
AbstractPlugin::configuration(), this allows to modify the configuration
even before the plugin is loaded (but it has to be known) and it also
affects all plugins loaded later through the same manager. That's useful
for example to affect JPEG output quality for image conversion plugins
that are loaded through AnyImageConverter -- i.e., there's no other way
to configure anything in a live instance because they are loaded,
instantiated and used in a single step internally.

In order to make the global configuration testable and predictable also
for static plugins, I had to change the internals to parse the
configuration and metadata in manager constructor when it acquires
unused static plugins from the global storage instead of doing it during
static plugin registration. The manager then clears the metadata on
destruction and so it's in the same initial state the next time a
manager instance is created.

The metadata is now an Optional, meaning it will blow up in case it's
accessed in a detached state. The test passes so I hope it won't cause
regressions in user code.