Drop support for CMake 2.8.12.
Also finally. Ubuntu 14.04 is way too old and it doesn't make sense
supporting it anymore (even Travis provides 16.04 now). Unfortunately
Debian 8 (Jessie) is still on 3.0.2, so that was meant to be the minimal
version. However that version has a bug which doesn't correctly
propagate local imported targets upwards, so we have to use 3.1 instead.

Removed also all CMake policy settings that are implicitly enabled for
3.1 (which is, lucklily, all of them).