Utility: move STL-specific stuff to DebugStl.h.
There's still <iosfwd> and <utility> lurking around, but even with that
the amount of preprocessed lines goes down from 14k to 3k under C++11
and from 17k to 4k under C++17. Nice.

I had to use the IsStringLike type trait to avoid forward-declaring /
including std::string and due to that there's a new operator<< for
std::basic_string of types other than char, since those would be now
interpreted as string-like. For now I'm keeping the current behavior of
printing only std::string as a string and all other (e.g.
std::u32string) as containers, but maybe it might make sense to revisit
this in the future and print all those types as strings.