Interconnect: explicitly disable /OPT:ICF on MSVC.
Too many problems and inventing workarounds that make the code "look
different" to the compiler are absolutely not feasible. It was sorta
bearable in the library code but I can't force the users to do the same
in their code.

For CMake 3.13 and newer this uses INTERFACE target_link_options(), for
older versions this (ab)uses a hidden feature of target_link_libraries()
where passing a library with `-` in front causes it to be treated as a
flag and not a library. Fortunately the linker accepts the -OPT:NOICF
form as well next to /OPT:NOICF, so all is great.

Also added an explicit DANGER BOX with this information for the poor
souls who have to use a custom buildsystem.