Interconnect: rework this thing.
Until now, the slot connections were a mess of heap-allocated classes
with virtual calls and even that wasn't really giving the flexibility
needed. Now it is all merged into a single type with:

 * About 16 + 8 bytes of storage to fit even the craziest member
   function pointers together with the receiver pointer.
 * The same storage reusable for free function pointers (which need
   just 8 bytes out of that)
 * The same 24 bytes usable for trivially copyable capturing lambdas.
 * And, finally, for larger function objects or non-trivially-copyable
   lambdas a possibility to store them on heap and call a destructor
   when the signal is removed.

To actually call the slot, a specialized lambda is saved into a
type-erased function pointer for each case, then during emit() being
cast back to its type and called. This adds another 8 bytes, making it
32 bytes in total.

Finally, the original helpers for re-entrant signal emitting and a
connection type enum add another 5 bytes, rounding up to 40 bytes. The
original connection data kept a pointer to the Connection object and
the Emitter as well. That's no longer done to simplify the internals,
which resulted in minor breaking API changes in the Connection object.

As per my investigation the additional indirect call through a function
pointer is for some reason actually slower than the original
virtual-function-based implementation in Debug mode, however due to
cleanup in other parts the total performance increases. For the record,
current benchmark output is:

 BENCH [01] 298.35 ± 36.35  ns connect1kFunctions()@99x1000 (wall time)
 BENCH [02] 337.05 ± 25.29  ns connect1kMembers()@99x1000 (wall time)
 BENCH [03]   0.29 ± 1.83   µs destructBaseline()@99x1 (wall time)
 BENCH [04]  63.35 ± 8.41   µs destruct1kFunctions()@99x1 (wall time)
 BENCH [05]   4.05 ± 0.12   ms destruct1kMembersEmitterFirst()@99x1 (wall time)
 BENCH [06]  31.81 ± 0.51   ms destruct1kMembersReceiverFirst()@99x1 (wall time)
 BENCH [07]   1.66 ± 0.16   µs call1kFunctions()@24x100 (wall time)
 BENCH [08]  15.79 ± 1.11   µs call1kStdFunctions()@24x100 (wall time)
 BENCH [09]   6.32 ± 0.47   µs call1kFunctionConnectionData()@24x100 (wall time)
 BENCH [10]  10.53 ± 1.69   µs call1kLambdaConnectionData()@24x100 (wall time)
 BENCH [11]   9.71 ± 0.94   µs call1kLambdaHeapConnectionData()@24x100 (wall time)
 BENCH [12]  11.73 ± 0.93   µs call1kMemberConnectionData()@24x100 (wall time)
 BENCH [13] 120.44 ± 2.17   µs callSlotFunction1000x()@24x100 (wall time)
 BENCH [14]  69.46 ± 3.91   µs call1kSlotFunctions()@24x100 (wall time)
 BENCH [15]  67.31 ± 2.03   µs call1kSlotLambdas()@24x100 (wall time)
 BENCH [16]  67.54 ± 2.10   µs call1kSlotLambdasHeap()@24x100 (wall time)
 BENCH [17]  70.90 ± 1.54   µs call1kSlotMembers()@24x100 (wall time)

Out of those ~67 ms around 60 is the overhead of std::unordered_map,
which is absolutely crazy. This however has to wait until I have a
replacement structure. There are other optimization opportunities (such
as avoiding the extra indirection and calling free / member functions
directly if possible), but those are quite dangerous so I'm postponing
them to later.

Co-authored-by: Jakob Weiss <jakob.weiss@tum.de>