TestSuite: another, better look at diagnostic file saving.
Now the comparator can tell the tester if it has any file to save
(sometimes stuff fails *so bad* that there just isn't, and since we
can't fail in the save function, we need to handle that before). In
addition, the diagnostic file can be also saved even if the comparison
*didn't* fail. Because feature parity and all that. The option is thus
now called --save-diagnostic instead of --save-failed.

And because I felt like extra bikeshedding today, the comparator is
additionally able to print warnings and messages, and finally there's a
way to enable --verbose output, in which case comparators can print
verbose messages and save verbose diagnostic files as well.

This required a change to Comparator signatures. To keep old things
working, the Tester has a shim for calling the old signatures, but
that's (loudly) deprecated and will be removed in a future release.
Builtin comparators are also not updated yet, the commit would be too
big otherwise.