Utility: replace dllLocation() with multiplatform libraryLocation().
The dllLocation() API was there just for a while (not a part of a
versioned release, even), so I'm not doing any deprecation round
for it. Complain loudly if you started depending on it, I'll add
that back (but deprecated).

This also makes Utility depend on libdl on Unix platorms (before
it was just PluginManager). Not sure how to feel about that (also
not sure if that can be worked around).