modules: avoid plugin DLLs being relinked every time.
The BYPRODUCTS path was wrong and so Ninja futilely rebuilt the target
every time in a hope that the path would then exist. However I can't use
a generator expression there so just commenting that out for now. This
would break once something starts depending on the copied conf file but
I doubt that's a very wide use case.