Utility: add better Directory::map(), rename original to mapWrite().
The original one doesn't preserve file contents, so it's practically
write-only. Its original intended purpose was a "swap file" on the disk
and should have been named mapWrite() right from the start.

The new map() is like mapRead() but with read-write access and failing
when a file doesn't exist. The original map() signature is still
present, but marked as deprecated.