Utility: implement Duff's device for small copy() inner loops.
It is chosen either when the last dimension is less than 8 bytes or when
the last dimension is also non-contiguous, and on Clang only as GCC
behaves better with a simple loop. (Still need to check what's better on
MSVC.)

In the benchmark I saw a speedup from 17 microseconds to 10 for a
release build; it made the Debug build slightly worse, but still faster
than a plain loop so I think we're good.