Drop CORRADE_BIG_ENDIAN and expose CORRADE_TARGET_BIG_ENDIAN.
Similarly as the CORRADE_TARGET_X86 etc defines, it didn't make sense to
have this propagated from CMake as the value could be different on
platforms with multi-architecture binaries. Instead, it's now defined
inside configure.h based on platform defines (which I shamelessly stole
from SDL).

Also, to follow the naming of all other macros, it's renamed to
CORRADE_TARGET_BIG_ENDIAN, with the old name being temporarily kept for
backwards compatibility. Additionally, it seems that there was no code
that actually tested this macro is correct (we just relied on CMake
doing the right thing), so TargetTest was extended for that. And this
test will likely also fail on mixed-endian systems and Bi-Endian systems
that were switched to the other endianness at runtime.