Utility: added a CORRADE_LONG_DOUBLE_SAME_AS_DOUBLE macro.
Long overdue, the various inconsistent ifdefs in all places dealing with
floating-point precisions were a nightmare. Of course nothing is ever
simple, so this made me discover that Emscripten sometimes has
long doubles as 80-bit and sometimes as 64-bit. The smaller variant was
on 1.38.44 with Clang 6, while the larger on 1.38.38 with Clang 9.
Fortunately I quickly discover a preprocessor macro that could check for
this, so hopefully all good.