Containers: disallow creating void ArrayViews from const T.
It failed compilation before as well, but it could cause ambiguous
overload errors because it was not disabled via SFINAE directly in the
function signature.

Well, it still causes ambiguous overloads when passing e.g. Array<char>
to a function overload set that accepts either ArrayView<const void> or
ArrayView<void>, but even C++27 is too shitty to handle this in even a
remotely sane way without having to haul most of the code inline into a
header, so eh.