Utility: CORRADE_ASSERT_UNREACHABLE() with a message.
This is a breaking change -- the original macro is now available as
CORRADE_INTERNAL_ASSERT_UNREACHABLE(). This also makes it consistent
with other assert macros, so it isn't just some arbitrary breakage.

It was on my TODO list for quite a while, but now I ran into an issue
where I needed to compile with CORRADE_NO_ASSERT enabled and it was
suddenly not possible, as all `CORRADE_ASSERT(false, ...)` that
communicated the unreachability to the compiler got removed. This is now
also documented as a "no no" use case for CORRADE_ASSERT() /
CORRADE_INTERNAL_ASSERT().