PluginManager: provide a helper for reliable plugin search paths.
Having everything copied into thousands of places doesn't really help,
so finally unifying that. This isn't used yet in Corrade itself (it will
be once the filesystem abstraction starts happening). Compared to what
was done in Magnum before this makes the hardcoded option first again
in order to provide a reliable override in case everything else fails.

Additionally this adds an Apple-specific search path for bundles.