PluginManager: get rid of shared mutable global state.
So clean. So beautiful.

It was only needed for inter-manager dependencies which are not
currently used by Magnum at all (but they probably will be again in the
future). Because of that, I don't feel breaking this feature by
introducing a now-mandatory call to registerExternalManager() to
supplement the lack of global state.

In addition this gets rid of various tests that were only there to check
for stuff specific to the shared global state.