Utility: avoid string copy on return in Arguments and ConfigurationGroup.
This is needed in order to be able to return non-dangling string views
from there, but is a good thing to do in any case -- copying a string
just to return it converted to an integer is silly.

In case of Arguments we need to return some reference in case of
graceful asserts, so there's some extra work to do in the tests. But
usual user code doesn't have graceful asserts enabled, so this should be
harmless for the real user.