Containers: added String[View]::partition().
At first I wanted to add a new overload to Utility::String, but there
it's a hell of ambiguous overloads that I don't want to deal with. The
utility namespace primary reason was to extend the standard string
class, now we can simply add to our own class instead.

Another potential aspect why to keep own namespace is saving compilation
times when there will be many APIs, but I don't expect that to become a
problem -- all functions are deinlined and potentially complex return
types are forward-declared, so compared to the FULLY INLINE hell of
std::basic_string this is completely fine.

So, next I'm moving the new split() overloads from the String namespace
to the String[View] classes.