Containers: implement split() directly on String[View].
Because it doesn't inflate the headers in any way, and the
ambiguous overload hell in Utility::String was just too much pain to
handle. The original overloads in Utility::String wre around just for a
while, but I'm keeping (and deprecating) them instead of removing, since
it's not too much extra work.

New APIs will be probably implemented directly on the String[View]
classes, not added to Utility::String.