Containers: don't make ArrayTuple.h implicitly depend on ArrayView.h.
I want to use this class as a more beefed-up variant of Pointer for
PIMPLing class internals that need nested allocations, and the extra
header dependency (which is rather large, and drags in Assert.h, which
uses Debug.h...) was bothering me. Stats with g++ -E -P -std=c++XY:

    #include <header>              | C++11 | C++20
    -------------------------------+-------+------
    Pointer.h	                   | 4197  | 5943
    Pointer.h, NO_ASSERT, NO_DEBUG | 2451  | 4120
    ArrayTuple.h before            | 4875  | 6668
    ArrayTuple.h before, NO_ASSERT | 3177  | 4968
    ArrayTuple.h after             | 2872  | 4663

Because ArrayTuple is a non-templated type and doesn't have any publicly
used assertions, it's actually smaller than a Pointer (and not that much
larger than it w/o debug and assertions), which is nice.

I could probably consider moving the guts to a hpp file, saving even
more, but this seems good enough for now. Will revisit if it becomes a
bottleneck after heavy use.