Containers: renamed String::strip{Pre,Suf}fix() to except{Pre,Suf}fix().
Because the original naming confusingly made the following code look
like it should be doing something -- while it shouldn't:

    Containers::String a = "not confused!";

    a.stripPrefix("not "); // does not mutate, returns a view

Same for StringView as well. Even though the classes are
"relatively" new and not part of any release yet, the original APIs are
kept and marked as deprecated -- I don't want people to suffer
breakages.