Containers: test LinkedListItem::list() access during destruction.
If an item is a part of a list, it feels to me like it makes sense to be
able to access the list() inside its subclass destructor, but it
doesn't seem to be the case. This code was touched last in
ca4d12fbf277e0acebac0d566516acd2aa3f5d92 but before that the same
problem seems to be there -- first disconnecting it from the list and
only then calling the destructors.

Fix in the next commit.