Aaaaaand reenable dynamic plugins once more
They do work in dynamic configuration, most test checks pass, and
the ones that don't are primarily due to bazel being incapable
of providing full plugin path.
Static Corrade + dynamic plugins still has issues with bazel,
but it is an edge case in bazel world, and considering the way
bazel resolves those things, in the end it will fall into the
user's hands to deal with on a case by case basis anyway.