TestSuite: I got fed up with the parentheses in CORRADE_VERIFY().
((Why would I)) need to (((wrap all the things in a parentheses, I
don't))) understand. This way it's much better. So much
unnecessary suffering.

OTOH doing the same for CORRADE_COMPARE() could be harmful, so don't,
and instead add a comment to prevent myself from accidentally trying
that in the future.