Containers: drop const && getter overloads from MoveReference.
These cause ambiguous overload errors on GCC 4.8 which I would need to
disable there. As such these would cause unnecessary extra trouble in
all code using MoveReference, because the behavior will be always
slightly different on this compiler and I would need to remember to call
`std::move(*a)` instead of `*std::move(a)` and similar minor
differences.

Because this type is extremely specific and I don't expect it to be used
much (if at all) outside of Magnum libraries themselves, I don't see
such deliberately missing APIs as a problem right now. Besides that, as
the Reference is a "view" type, the getters were overloads on `const &&`
and `const &`, which alone is crazy weird and unheard of in common C++
code.