Utility: enable CORRADE_FALLTHROUGH on MSVC 16.6 under C++17.
Both MSVC and Clang require the [[fallthrough]] to be used only on
C++17, unfortunately there's nothing pre-C++17 on MSVC so we enable this
macro only on C++17 and for Clang we use the old thing always.

OTOH we can make GCC use the C++11 attribute syntax for this -- it works
that way since GCC 7 where this attribute got introduced and doesn't
print any warnings.