Move the initialization and creation tags to the root namespace.
This was done for Magnum already and it's just a lot of unnecessary
typing. Given that the containers may eventually get moved there as
well, this makes sense.

Unfortunately to avoid the deprecation warnings I had to prefix all
current uses in the Containers library by Corrade::. Ugly, but temporary
only since the aliases get removed.