TestSuite: change CORRADE_FAIL() to CORRADE_FAIL_IF().
I just realized this API is bad to use and had to change it. No
backwards compatibility provided, since it's just about a month since
the original got introduced in d441e1962b90aa21993dbcb25b1e15c4af2aea8c.
Consider the following:

    void Test::foo() {
        if(a < 5)
            CORRADE_FAIL("We didn't expect to be so low.");
        if(a % 4 != 0)
            CORRADE_FAIL("It has to be aligned to four bytes.");
    }

If neither of those fails, the test will not hit any of the macros, and
thus gets reported as a failure due to not having any checks. To
mitigate that, one would have to add an explicit CORRADE_VERIFY(true)
at the end, but in tests with divergent code paths such practice would
only lead to false negatives and misery.

With the new macro, the above would be instead

    void Test::foo() {
        CORRADE_FAIL_IF(a < 5, "We didn't expect to be so low.");
        CORRADE_FAIL_IF(a % 4 != 0, "It has to be aligned to four bytes.");
    }

and there's clearly no chance of these macros not being called by
accident. If it's still desired to fail unconditionally, one can always
do a

    CORRADE_FAIL_IF(false, "Everything sucks!");

but I don't think such case is very common, so I won't provide a
dedicated CORRADE_FAIL() macro for this -- also, having it would
encourage the undesired practice above, and that we want to prevent.